requirements.txt
astroid==2.3.3
asyncio==3.4.3
Click==7.0
colorama==0.4.3
fastapi==0.65.2
websocket-client==1.4.1
gunicorn==20.0.4
h11==0.9.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pydantic==1.6.2
pylint==2.4.4
six==1.14.0
starlette==0.14.2
uvicorn==0.18.3
websockets==10.3
wrapt==1.11.2
uvloop==0.14.0
requests==2.27.1
pyjwt==2.5.0
python_jwt==4.0.0
paho-mqtt==1.6.1
cryptography==37.0.4
jwcrypto==1.4.2